# Speed up `gem install <gem-that-does-not-exist>`