@deivid-rodriguez 2023-09-25
I'd like to refactor things to not have to match exceptions by message. It doesn't seem trivial though.