@deivid-rodriguez 2023-09-25
As noticed by @simi at https://github.com/rubygems/rubygems/issues/6821#issuecomment-1733356709, this may be completing removing the API fallback from `gem install` which is not what we want. So I set it as a draft.