@deivid-rodriguez 2023-10-19
I don't think this is the right solution, so let me close this!

I think the best approach would be what we discussed at https://github.com/rubygems/rubygems/issues/6821#issuecomment-1733378708 and follow up comments.